requirements.txt
blinker==1.9.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
colorzero==1.1
dbus-python==1.2.16
distro==1.5.0
distro-info==1.0+deb11u1
flask==3.1.2
gpiozero==1.6.2
idna==2.10
importlib-metadata==8.7.0
influxdb-client==1.49.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
paho-mqtt==2.1.0
pycurl==7.43.0.6
PyGObject==3.38.0
python-apt==2.2.1
python-dateutil==2.9.0.post0
reactivex==4.0.4
requests==2.25.1
RPi.GPIO==0.7.0
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
typing-extensions==4.15.0
unattended-upgrades==0.1
urllib3==1.26.5
werkzeug==3.1.3
zipp==3.23.0